AP-14139: Introduce RowContainer interface and Buffered implementation
just legacy code
